+++ src/main/java/co/invest72/exchange_rate/infrastructure/api/FixedExchangeRateProvider.java
@@ -1,14 +1,11 @@
 package co.invest72.exchange_rate.infrastructure.api;
 
-import java.math.BigDecimal;
 import java.util.Objects;
-import java.util.Optional;
 
 import co.invest72.exchange_rate.application.ExchangeRateUpdateHandler;
 import co.invest72.exchange_rate.domain.Currency;
 import co.invest72.exchange_rate.domain.ExchangeRateProvider;
 import co.invest72.exchange_rate.domain.service.ExchangeRateService;
-import co.invest72.money.domain.Pair;
 import lombok.extern.slf4j.Slf4j;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
@@ -22,11 +19,6 @@ public class FixedExchangeRateProvider implements ExchangeRateProvider {
 		ExchangeRateUpdateHandler exchangeRateUpdateHandler) {
 		this.exchangeRateService = Objects.requireNonNull(exchangeRateService);
 		this.exchangeRateUpdateHandler = Objects.requireNonNull(exchangeRateUpdateHandler);
-	}
-
-	@Override
-	public Optional<BigDecimal> getRate(Currency from, Currency to) {
-		return exchangeRateService.getRate(new Pair(from, to));
 	}
 
 	@Override
